# internal/models/subscription.go
package models

import (
	"github.com/google/uuid"
	"time"
)

type Subscription struct {
	ID          uuid.UUID  `json:"id" db:"id"`
	ServiceName string     `json:"service_name" db:"service_name" binding:"required"`
	Price       int        `json:"price" db:"price" binding:"required,gte=0"`
	UserID      uuid.UUID  `json:"user_id" db:"user_id" binding:"required"`
	StartDate   time.Time  `json:"start_date" db:"start_date" binding:"required"`
	EndDate     *time.Time `json:"end_date,omitempty" db:"end_date"`
	CreatedAt   time.Time  `json:"created_at" db:"created_at"`
	UpdatedAt   time.Time  `json:"updated_at" db:"updated_at"`
}

type CreateSubscriptionRequest struct {
	ServiceName string    `json:"service_name" binding:"required"`
	Price       int       `json:"price" binding:"required,gte=0"`
	UserID      uuid.UUID `json:"user_id" binding:"required"`
	StartDate   string    `json:"start_date" binding:"required"`
	EndDate     string    `json:"end_date,omitempty"`
}

type UpdateSubscriptionRequest struct {
	ServiceName string  `json:"service_name,omitempty"`
	Price       *int    `json:"price,omitempty"`
	StartDate   string  `json:"start_date,omitempty"`
	EndDate     *string `json:"end_date,omitempty"`
}

type SubscriptionFilter struct {
	UserID      uuid.UUID `form:"user_id"`
	ServiceName string    `form:"service_name"`
	StartDate   string    `form:"start_date"`
	EndDate     string    `form:"end_date"`
}

type TotalCostResponse struct {
	TotalCost int `json:"total_cost"`
}